test: separate tests for experimental api
